requirements.txt
# absl-py==1.1.0
# asgiref==3.5.2
# astunparse==1.6.3
# cachetools==5.2.0
# certifi==2022.5.18.1
# charset-normalizer==2.0.12
# colorama==0.4.4
# commonmark==0.9.1
# cycler==0.11.0
Django>=2.1.3,<2.2.0
djangorestframework==3.9.4
# flatbuffers==1.12
# fonttools==4.33.3
# gast==0.4.0
# google-auth==2.6.6
# google-auth-oauthlib==0.4.6
# google-pasta==0.2.0
# grpcio==1.46.3
# h5py==3.7.0
# idna==3.3
# imageio==2.19.3
# importlib-metadata==4.11.4
# keras==2.9.0
# Keras-Preprocessing==1.1.2
# kiwisolver==1.4.2
# libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
# networkx==2.6.3
numpy==1.21.6
# oauthlib==3.2.0
# opt-einsum==3.3.0
# packaging==21.3
# Pillow==9.1.1
# protobuf==3.19.4
# prototf==0.0.1
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# Pygments==2.13.0
# pyparsing==3.0.9
# python-dateutil==2.8.2
# pytz==2022.6
# PyWavelets==1.3.0
requests==2.27.1
# requests-oauthlib==1.3.1
# rich==12.5.1
# rsa==4.8
# scikit-image==0.19.2
# scipy==1.7.3
# six==1.16.0
# sqlparse==0.4.3
# tensorboard==2.9.0
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
# tensorflow-estimator==2.9.0
# tensorflow-io-gcs-filesystem==0.26.0
# termcolor==1.1.0
# tifffile==2021.11.2
# tqdm==4.64.0
# typing_extensions==4.2.0
# tzdata==2022.6
# urllib3==1.26.9
# Werkzeug==2.1.2
# wrapt==1.14.1
# zipp==3.8.0
python-dotenv